requirements.txt
app==0.0.1
boto3==1.37.36
fastapi==0.115.12
Pillow
torch
torchvision==0.21.0
uvicorn[standard]
python-multipart
pytest
